Update Name Resolution #863
This puts in place a temporary fix for name resolution which allows
packages to store WyIL files in arbitrary places, rather than rooted at
the root of the package (i.e. zip file).  This makes a lot of sense in
the long run.  But, the changes are really a hack.  WhileyFileResolver
in general is now a completely broken concept.

One potential issue with this fix is that it is degrading performance I
believe.  The problem is that performance is now dependent on the number
of wyil files in the given directory.  In most cases, this wouldn't be a
problem but for the test cases where you have 500+ wyil files it is a
real problem.